Código actualizado.
> Se corrigió el modelo Cliente-Servidor : El cliente ahora es quien recibe de forma activa los datos del puntero del servidor.
