+++ src/Cliente/Cliente.java
@@ -2,26 +2,37 @@ package Cliente;
 
 import Utillities.Validaciones;
 
-import java.awt.*;
 import java.io.IOException;
-import java.net.*;
-
-import static java.lang.Thread.sleep;
+import java.net.DatagramPacket;
+import java.net.DatagramSocket;
+import java.net.SocketException;
 
 public class Cliente {
 
+    private byte[] MsjRecibidoBytes;
     private DatagramSocket Socket;
-    private DatagramPacket Packet;
-    private InetAddress Dirección;
-    private String Mensaje;
-    private byte[] MensajeBytes;
-    private int Puerto;
+    private DatagramPacket PaqueteRecibido;
+    private String MsjRecibido;
+
+
+    /**
+     * Para inicializar la instancia del objeto desde su respectivo launcher.
+     */
+    public void iniciar() {
+        try {
+            inicializarDatos();
+            menuPrincipal();
+        } catch (SocketException e) {
+            System.err.println(e.getMessage());
+        }
+    }
+
 
     /**
      * Inicia el menú principal. Este método será llamado por su respectivo
      * launcher en el paquete.
      */
-    public void menuPrincipal() {
+    private void menuPrincipal() {
         opcionesMenuPrincipal();
         seleccionMenuPrincipal();
     }
@@ -31,10 +42,11 @@ public class Cliente {
      */
     private void opcionesMenuPrincipal() {
         System.out.println("""
+                
                 [ICC 260-1]
                 
                 ¿Qué deseas hacer?
-                [1] Mandar datos del puntero al servidor.
+                [1] Obtener datos del puntero en el servidor.
                 [2] Salir.
                 """);
     }
@@ -46,7 +58,7 @@ public class Cliente {
         boolean quedarse = true;
 
         switch (Validaciones.validarIntervalo(1, 2)) {
-            case 1 -> enviarDatos();
+            case 1 -> abrirEntrada();
             case 2 -> quedarse = false;
         }
 
@@ -59,36 +71,78 @@ public class Cliente {
     }
 
     /**
-     * Envía los datos al servidor, y pide la cantidad de segundos para
-     * enviarlos.
+     * Permite la entrada de datagramas por un número determinado de segundos.
      */
-    private void enviarDatos() {
+    private void abrirEntrada(){
+        int tiempo = preguntarTiempo();
+        ejecutar(tiempo);
+    }
+
+    /**
+     * Ejecuta el programa de forma constante, y actúa en caso de que llegue
+     * un paquete.
+     * En caso de excepción, el programa se detiene y muestra los datos
+     * asociados al error.
+     */
+    public void ejecutar(int segundos) {
         try {
-            conectar();
-            int tiempo = preguntarTiempo();
-            enviarRáfagaPaquetes(tiempo);
-            System.out.println("[Fin de la transmisión]");
+            long tiempoEjecucion = System.currentTimeMillis() + segundos * 1000L;
+            while (System.currentTimeMillis() <= tiempoEjecucion) {
+                recibirPaquete();
+                procesarMensaje();
+            }
         } catch (Exception e) {
-            System.err.println(e.getLocalizedMessage());
+            System.err.println(e.getClass());
+            System.err.println(e.getMessage());
         }
     }
 
+    /**
+     * Inicializa los datos con los cuales se trabajará.
+     * > El puerto a utilizar para cliente y servidor será el '54321'.
+     * > El largo máximo en bytes que pueden tener los mensajes es de 256.
+     * @throws SocketException
+     */
+    private void inicializarDatos() throws SocketException {
+        this.MsjRecibidoBytes = new byte[256];
+        this.Socket = new DatagramSocket(54321);
+        this.PaqueteRecibido = new DatagramPacket(this.MsjRecibidoBytes,256);
+
+    }
 
     /**
-     * Establece parámetros de conexión para el programa.
-     * @throws SocketException
-     * @throws UnknownHostException
+     * Recibe el paquete a través del socket.
+     * @throws IOException
      */
-    private void conectar() throws SocketException, UnknownHostException {
+    private void recibirPaquete() throws IOException {
+        this.Socket.receive(this.PaqueteRecibido);
+    }
+
+    /**
+     * Pasa el mensaje a variable global y lo parametriza.
+     */
+    private void procesarMensaje() {
         try {
-            this.Socket = new DatagramSocket();
-            this.Dirección = InetAddress.getByName("localhost");
-            this.Puerto = 3333;
-        } catch (SocketException e) {
-            throw new SocketException("El socket no es válido");
-        } catch (UnknownHostException e) {
-            throw new UnknownHostException("Host desconocido");
+            this.MsjRecibido = new String(this.MsjRecibidoBytes).trim();
+            parametrizarMensaje(this.MsjRecibido);
+        } catch (Exception e) {
+            System.err.println(e.getMessage());
         }
+    }
+
+    /**
+     * Desde el cliente, el formato del mensaje era 'x;y;seg', este método
+     * toma los valores, los separa, y los muestra de forma más sofisticada.
+     * @param mensaje
+     * @throws Exception
+     */
+    private void parametrizarMensaje(String mensaje) throws Exception {
+        String[] partesMensaje = mensaje.split(";");
+        if (partesMensaje.length !=2) {
+            throw new Exception("El paquete no cumple con parámetros");
+        }
+        System.out.println("\nPosición horizontal: " + partesMensaje[0]
+                        + "\nPosición vertical: " + partesMensaje[1]);
     }
 
     /**
@@ -102,98 +156,6 @@ public class Cliente {
                 de tu puntero? (Número entero).
                 """);
         return Validaciones.validarPositivo();
-    }
-
-    /**
-     * Envía una ráfaga de mensajes cada 0.1 segundos durante la cantidad de
-     * segundos ingresadas como parámetro.
-     * @param segundos
-     * @throws IOException
-     * @throws InterruptedException
-     */
-    private void enviarRáfagaPaquetes(int segundos) throws IOException,
-            InterruptedException {
-        try {
-            for (float ms = 0 ; ms <=  segundos; ms += 0.2) {
-                Point ubicaciónPuntero = obtenerUbicacionPuntero();
-
-                double[] coordenadasPuntero = extraerCoordenadas(ubicaciónPuntero);
-                double x = coordenadasPuntero[0];
-                double y = coordenadasPuntero[1];
-
-                String datosPaquete = parametrizarDatos(x, y, ms);
-                generarMensaje(datosPaquete);
-                enviarMensaje();
-
-                sleep(200);
-            }
-        } catch (InterruptedException e) {
-            throw new InterruptedException("Algo interrumpió el proceso.");
-        }
-    }
-
-    /**
-     * Envía un paquete de datagrama a la dirección localhost, y al puerto
-     * asignado en la función conectar(), el 3333.
-     * @throws IOException
-     */
-    private void enviarMensaje() throws IOException {
-        try {
-            if (this.MensajeBytes.length > 256) {
-                throw new IOException("El paquete es de más de 256 bytes.");
-            }
-            this.Packet = new DatagramPacket(this.MensajeBytes,
-                    this.MensajeBytes.length,
-                    this.Dirección,
-                    this.Puerto);
-            this.Socket.send(this.Packet);
-        } catch (Exception e) {
-            throw new IOException("Error al enviar el paquete.");
-        }
-
-    }
-
-    /**
-     * Dada una string, establece los valores de las variables globales de
-     * Mensaje y su equivalente en bytes, MensajeBytes.
-     * @param mensaje
-     */
-    private void generarMensaje(String mensaje) {
-        this.Mensaje = mensaje;
-        this.MensajeBytes = mensaje.getBytes();
-    }
-
-    /**
-     * Para enviar un paquete, los datos se deben serializar, es decir, pasar
-     * los datos a String para encajar con los parámetros.
-     * @param x
-     * @param y
-     * @param tiempo
-     * @return
-     */
-    private String parametrizarDatos(double x, double y, float tiempo) {
-        return x + ";" + y + ";" + tiempo;
-    }
-
-    /**
-     * Dada la ubicación del puntero, extrae las coordenadas en X e Y, y
-     * las pasa a un arreglo d <double>.
-     * @param puntero
-     * @return
-     */
-    private double[] extraerCoordenadas(Point puntero) {
-        double x = puntero.getX();
-        double y = puntero.getY();
-        return new double[]{x, y};
-    }
-
-    /**
-     * Obtiene la ubicación actual del puntero (Mouse).
-     * @return
-     */
-    private Point obtenerUbicacionPuntero() {
-        PointerInfo ubicaciónPuntero = MouseInfo.getPointerInfo();
-        return ubicaciónPuntero.getLocation();
     }
 
 }
