+++ src/Cliente/LauncherCliente.java
@@ -1,0 +1,9 @@
+package Cliente;
+
+public class LauncherCliente {
+
+    public static void main(String[] args) {
+        Cliente cliente = new Cliente();
+        cliente.iniciar();
+    }
+}
