+++ src/Servidor/LauncherServidor.java
@@ -4,7 +4,7 @@ public class LauncherServidor {
 
     public static void main(String[] args) {
         Servidor servidor = new Servidor();
-        servidor.ejecutar();
+        servidor.iniciar();
     }
 
 }
